+++ packages/builder-vue3/src/hooks/resolve-dir-import.ts
@@ -5,16 +5,25 @@ import { resolveDirImport } from '@varlinor/node-tools/plugins'
  * 可以将scripts中的 目录应用补全。以防二次引用时路径解析错误。
  * @param packageRoot
  * @param scopes
+ * @param exts
+ * @param subPackageBase
  * @returns
  */
-export function rewriteDirImportPath(packageRoot: string, scopes: string[]): function {
+export function createRewriteHook(
+  packageRoot: string,
+  scopes: string[],
+  exts?: string[],
+  subPackageBase?: string = 'packages/'
+): function {
   return (ctx, opts) => {
     // console.log('current opts:', opts)
     const { plugins } = opts
     if (Array.isArray(plugins)) {
       const plugin = resolveDirImport({
         basePath: packageRoot,
-        scopes
+        exts,
+        scopes,
+        subPackageBase
       })
 
       const idx = plugins.findIndex((p) => {
@@ -26,4 +35,27 @@ export function rewriteDirImportPath(packageRoot: string, scopes: string[]): fun
       }
     }
   }
+}
+
+/**
+ * 用于 vite 的 plugin
+ * 在打包scripts时，添加resolve dir import 插件使用。
+ * @param packageRoot
+ * @param scopes
+ * @param exts
+ * @param subPackageBase
+ * @returns
+ */
+export function createRewritePlugin(
+  packageRoot: string,
+  scopes: string[],
+  exts?: string[],
+  subPackageBase?: string = 'packages/'
+): any {
+  return resolveDirImport({
+    basePath: packageRoot,
+    exts,
+    scopes,
+    subPackageBase
+  })
 }