+++ packages/builder-vue3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @varlinor/builder-vue3
 
+## 2.0.2
+
+### Patch Changes
+
+- feat: support to create hook for unbuild and plugin for vite.
+- Updated dependencies
+  - @varlinor/node-tools@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes